First basic test for data management view and moved common test code
